Change class visibility, refactor methods and add tests
Changed the ReturningError class to internal for encapsulation. Refactored methods in Extension.cs and ReturningGeneric.cs to use ReturningBase instead of Returning for cleaner inheritance and flexibility. Commented out a block of code in ReturningGeneric.cs - the Try method was removed due to its redundancy and code inefficiency. Added tests in ReturnValueOKTest.cs to ensure code changes did not break functionality. These changes improve the readability, efficiency, and testability of the code.

Took 2 hours 55 minutes