Update project versions
Updated the package versions in `MgSoftDev.Returning` and `MgSoftDev.Returning.Serilog` projects. The version number has been incremented to reflect the recent changes and improvements in the application. In addition, updated the `PackageReference` to the latest version of `MgSoftDev.Returning` and `Serilog`. Upgrades are essential to enhance the functionality and performance of the project.

Took 10 minutes